egl: use eglGetProcAddress instead of EGL_EGLEXT_PROTOTYPES
eglCreateImageKHR, eglDestroyImageKHR, and glEGLImageTargetTexture2DOES
are not defined in the symbol table in newer versions (>=10.6? or
possibly earlier) of the mesa libEGL and libGLESv2 libraries.  They
should be looked up at runtime instead.  Thus, look them up with the
eglGetProcAddress at runtime.  The use of EGL_EGLEXT_PROTOTYPES is no
longer needed for this.

Fixes #229

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>