encoder/jpeg: use new JPEG constructs
Use the new JPEG constructs from
codecparsers/jpegParser.[h|cpp] and
remove the old jpegparser.[h|c].

Add JPEG encode raw data unit tests

Test simple 10x10 NV12, I420, and YUY2 encoding to JPEG
format.

On appropriate HW, these tests currently fail because the
encoder does not report the correct max output size to the
client for sizes < 643 (w*h*3/2).  This needs to be fixed
in the encoder.

Currently, we are unable to attempt testing other raw fourcc
data since the encoder forces the test program to abort
(via an ASSERT in createSurface).  The encoder should not
assert, but rather return an error to the caller to
handle it appropriately; this needs fixed in the encoder,
too.

As expected, these tests will fail on HW that does not
support the VAAPI VAEntrypointEncPicture for JPEG.  These
tests can be filtered from the command line before
execution in this case.

Change-Id: Ieedd3c8d475716d231dc27d36f890eb864955a70
Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>
Signed-off-by: jkyu <jiankang.yu@intel.com>