jpegdec: we should not use YamiStatus as bool
In original define, the DECODE_SUCCESS is 1. All unit test will passed.
New writed code we define YAMI_SUCCESS as 0, and use it replace DECODE_SUCCES
We'd better not count on it's value and compare it to bool.
We need compare return value with DECODE_SUCCESS explicitly.

this will fix regression introduce in this patch

     commit 62ff8d9869ba6d69d5a368b534aca651bbe15aac
     Author: Xu Guangxin <guangxin.xu@intel.com>
     Date:   Tue May 31 16:14:04 2016 +0800

         decoder,encoder: use YamiStatus instead of Decoder_Status and Encoder_Status

         Change-Id: I07f2fcd64cd8993e06394f0e3e0e768089823da2
         Signed-off-by: Xu Guangxin <guangxin.xu@intel.com>

Change-Id: Ifef2bc8a1f22edb70364f538f59f9e652cad7b2e
Signed-off-by: Xu Guangxin <guangxin.xu@intel.com>