vpp: fix wrong header guard in oclpostprocess_base.h
The header inclusion guard was using vaapipostprocess_base_h
which conflicts with the real one found in vaapipostprocess_base.h
header file.  This is a bug and causes problems if
both "base" files cross paths in the same compilation unit.

Thus, use the correct name for the inclusion guard in
oclpostprocess_base.h header.

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>