decode/unittest_host: add host factory tests
Tests for the vaapidecoder_host functions to ensure the
decoders are properly registered with the factory.

The vaapidecoder_host_unittest is separated into its own
unittest_host executable so that we can detect static
initialization bugs with the decoder factory.  Separation
is required since any derived decoder that is explicitly
constructed in another test would hide such bugs.

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>