encode/unittest_host: add host factory tests
Tests for the vaapiencoder_host functions to ensure the
encoders are properly registered with the factory.

The vaapiencoder_host_unittest is separated into its own
unittest_host executable so that we can detect static
initialization bugs with the encoder factory.  Separation
is required since any derived encoder that is explicitly
constructed in another test would hide such bugs.

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>