h265dec: use int32 instead of uint16 for poc calculation
According to spec 7.4.3.2.1, the value of log2_max_pic_order_cnt_lsb_minus4
shall be [0, 12], so MaxPicOrderCntLsb can be 0x10000, out of the range of
uint16_t;
For m_prevPicOrderCntMsb, according to spec 8.3.1, is set value as following:
    picOrderCntMsb = m_prevPicOrderCntMsb + MaxPicOrderCntLsb;
    m_prevPicOrderCntMsb = picOrderCntMsb;
so m_prevPicOrderCntMsb may be greater than MaxPicOrderCntLsb(0x10000),
so use int32_t for it.

For example, there is a clip whose m_prevPicOrderCntMsb is actually equal
to 0x10000.

Signed-off-by: wudping <dongpingx.wu@intel.com>